+++ projects/ngx-input-counter/src/lib/ngx-input-counter.module.ts
@@ -1,17 +1,11 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MinusContentDirective, NgxInputCounterComponent, PlusContentDirective } from './ngx-input-counter.component';
 
-
-
 @NgModule({
-  declarations: [
+  imports: [
     NgxInputCounterComponent,
     PlusContentDirective,
     MinusContentDirective,
-  ],
-  imports: [
-    CommonModule
   ],
   exports: [
     NgxInputCounterComponent,