+++ projects/ngx-input-counter/src/lib/ngx-input-counter.service.ts
@@ -1,9 +1,22 @@
-import { Injectable } from '@angular/core';
+import { Injectable, TemplateRef, Type } from '@angular/core';
+import { TemplateContext } from './ngx-input-counter.component';
 
 @Injectable({
   providedIn: 'root'
 })
 export class NgxInputCounterService {
-
-  constructor() { }
+  value?: number;
+  name?: string;
+  class?: string;
+  min?: number;
+  max?: number;
+  step?: number;
+  minusTemplate?: TemplateRef<TemplateContext>;
+  plusTemplate?: TemplateRef<TemplateContext>;
+  plusComponent?: Type<any>;
+  minusComponent?: Type<any>;
+  minusClass?: string;
+  plusClass?: string;
+  valueClass?: string;
+  disabled?: boolean;
 }